Change default subscription mode to not return an initial table snapshot.
So far differential queries first returned an initial snapshot of the
table, and then ongoing diffs from there onwards. Because doing so
leads to substantial event volume when connecting to an agent, the
default now skips the initial snapshot: the first result will just be
empty, and it's then diffs from there on. The old behavior can be
returned through a new subscription type
`ZeekAgent::SnapshotPlusDifferences`. All the table scripts gain an
option to change the type they are using.